Generate models from JSON | Generate all models for data modeler
