+++ template/app/generated_config/models/organisations/models.json
@@ -1,6 +1,6 @@
 {
     "organisations":{ 
         "model": "Organisations",
-        "fields": {"organisation_name": {"dataType": "String(256)", "nullable": False, "unique": True, "relationship": None, "default": False}, "organisation_contact_name": {"dataType": "String(256)", "nullable": False, "unique": False, "relationship": None, "default": False}, "organisation_details": {"dataType": "Text", "nullable": True, "unique": False, "relationship": None, "default": False}, "organisation_contact_email": {"dataType": "String(256)", "nullable": False, "unique": False, "relationship": None, "default": False}, "organisation_address": {"dataType": "Text", "nullable": True, "unique": False, "relationship": None, "default": False}, "organisation_city": {"dataType": "String(256)", "nullable": True, "unique": False, "relationship": None, "default": False}, "organisation_postal_code": {"dataType": "String(256)", "nullable": True, "unique": False, "relationship": None, "default": False}, "organisation_country": {"dataType": "String(256)", "nullable": True, "unique": False, "relationship": None, "default": False}, "organisation_homepage": {"dataType": "Text", "nullable": True, "unique": False, "relationship": None, "default": False}, "organisation_vat_number": {"dataType": "Text", "nullable": True, "unique": False, "relationship": None, "default": False}, "organisation_reg_number": {"dataType": "Text", "nullable": True, "unique": False, "relationship": None, "default": False}}
+        "fields": {"organisation_name": {"dataType": "String(256)", "nullable": true, "unique": true, "relationship": null, "default": true}, "organisation_contact_name": {"dataType": "String(256)", "nullable": true, "unique": true, "relationship": null, "default": true}, "organisation_details": {"dataType": "Text", "nullable": true, "unique": true, "relationship": null, "default": true}, "organisation_contact_email": {"dataType": "String(256)", "nullable": true, "unique": true, "relationship": null, "default": true}, "organisation_address": {"dataType": "Text", "nullable": true, "unique": true, "relationship": null, "default": true}, "organisation_city": {"dataType": "String(256)", "nullable": true, "unique": true, "relationship": null, "default": true}, "organisation_postal_code": {"dataType": "String(256)", "nullable": true, "unique": true, "relationship": null, "default": true}, "organisation_country": {"dataType": "String(256)", "nullable": true, "unique": true, "relationship": null, "default": true}, "organisation_homepage": {"dataType": "Text", "nullable": true, "unique": true, "relationship": null, "default": true}, "organisation_vat_number": {"dataType": "Text", "nullable": true, "unique": true, "relationship": null, "default": true}, "organisation_reg_number": {"dataType": "Text", "nullable": true, "unique": true, "relationship": null, "default": true}}
     }
 }
