--- template/app/mod_test/templates/test/admin/index.html
@@ -1,124 +1,0 @@
-
-{% import 'macros/flash.html' as flash with context %}
-
-{% extends "admin/index.html" %}
-
-{% block content %}
-    
-    <div class="lg:flex lg:items-center lg:justify-between">
-        <div class="flex-1 min-w-0">
-            <h2 class="text-2xl font-bold leading-7 text-gray-900 sm:text-3xl sm:truncate">
-                Test
-            </h2>
-        </div>
-        <div class="mt-5 flex lg:mt-0 lg:ml-4">
-            <span class="sm:ml-3">
-                <a href="{{ url_for('test_admin.create') }}" class="inline-flex items-center px-4 py-2 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500">
-                Create
-                </a>
-            </span>
-        </div>
-    </div>
-    
-    <div class="hidden sm:block" aria-hidden="true">
-      <div class="py-5">
-        <div class="border-t border-gray-200"></div>
-      </div>
-    </div>
-
-    <!-- Flash Messages -->
-    {{ flash.render() }}
-    
-    <div class="flex flex-col">
-        <div class="-my-2 overflow-x-auto sm:-mx-6 lg:-mx-8">
-        <div class="py-2 align-middle inline-block min-w-full sm:px-6 lg:px-8">
-            <div class="shadow overflow-hidden border-b border-gray-200 sm:rounded-lg">
-            <table class="min-w-full divide-y divide-gray-200">
-                <thead class="bg-gray-50">
-                <tr>
-                    <th scope="col" class="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                        id
-                    </th>
-                    <!-- start new table headers -->
-
-                    <th scope="col" class="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                        Name
-                    </th>
-    
-                    <!-- end new table headers -->
-                    <th scope="col" class="relative px-6 py-3">
-                        <span class="sr-only">Edit</span>
-                    </th>
-                </tr>
-                </thead>
-                <tbody class="bg-white divide-y divide-gray-200">
-                {% for value in data.items %}
-                    <tr>
-                        <td scope="col" class="px-6 py-3 text-left text-xs font-medium text-gray-500 capitalize tracking-wider">
-                            {{ value.id }}
-                        </td>
-                        <!-- start new table values -->
-
-                        <td scope="col" class="px-6 py-3 text-left text-xs font-medium text-gray-500 capitalize tracking-wider">
-                            {{ value.name }}
-                        </td>
-    
-                        <!-- end new table values -->
-                        <td class="px-6 py-4 whitespace-nowrap text-right text-sm font-medium">
-                          <a href="{{ url_for('test_admin.edit', id=value.id) }}" class="text-indigo-600 hover:text-indigo-900">Edit</a> | 
-                          <a href="{{ url_for('test_admin.destroy', id=value.id) }}" class="text-red-600 hover:text-red-900">Delete</a>
-                        </td>
-                    </tr>
-                {% endfor %}
-                </tbody>
-            </table>
-                
-            <!-- Pagination Links-->
-            <nav class="relative z-0 inline-flex shadow-sm -space-x-px" aria-label="Pagination">
-                <a href="{{ url_for('test_admin.index', page=data.prev_num) }}"
-                class="relative inline-flex items-center px-2 py-2 rounded-l-md border border-gray-300 bg-white text-sm font-medium text-gray-500 hover:bg-gray-50
-                {% if data.page == 1 %}disabled{% endif %}">
-                <span class="sr-only">Previous</span>
-                <!-- Heroicon name: chevron-left -->
-                <svg class="h-5 w-5" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor" aria-hidden="true">
-                    <path fill-rule="evenodd" d="M12.707 5.293a1 1 0 010 1.414L9.414 10l3.293 3.293a1 1 0 01-1.414 1.414l-4-4a1 1 0 010-1.414l4-4a1 1 0 011.414 0z" clip-rule="evenodd" />
-                </svg>
-                </a>
-            <!-- Loop through the number of pages to display a link for each-->
-            {% for page_num in data.iter_pages(left_edge=1, right_edge=1, left_current=1, right_current=2) %}
-                {% if page_num %}
-                <!-- Check for the active page and set the link to "Active"-->
-                    {% if data.page == page_num %}
-                    <a href="{{ url_for('test_admin.index', page=page_num) }}"
-                    class="relative inline-flex items-center px-4 py-2 border border-gray-300 text-sm font-medium text-white bg-indigo-600">
-                        {{ page_num }}
-                    </a>
-                    {% else %}
-                    <a href="{{ url_for('test_admin.index', page=page_num) }}"
-                    class="relative inline-flex items-center px-4 py-2 border border-gray-300 bg-white text-sm font-medium text-gray-700 hover:bg-gray-50">
-                        {{ page_num }}
-                    </a>
-                    {% endif %}
-                {% else %}
-                {% endif %}
-            {% endfor %}
-            <a href="{{ url_for('test_admin.index', page=data.next_num) }}"
-            class="relative inline-flex items-center px-2 py-2 rounded-r-md border border-gray-300 bg-white text-sm font-medium text-gray-500 hover:bg-gray-50
-            {% if data.page == data.pages %}disabled{% endif %}">
-                <span class="sr-only">Next</span>
-                <!-- Heroicon name: chevron-right -->
-                <svg class="h-5 w-5" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor" aria-hidden="true">
-                <path fill-rule="evenodd" d="M7.293 14.707a1 1 0 010-1.414L10.586 10 7.293 6.707a1 1 0 011.414-1.414l4 4a1 1 0 010 1.414l-4 4a1 1 0 01-1.414 0z" clip-rule="evenodd" />
-                </svg>
-            </a>
-            </nav>
-            <p class="text-right mt-3">
-                <span class="text-gray-300">Showing page {{ data.page }} of {{ data.pages }}</span>
-            </p>
-
-            </div>
-        </div>
-        </div>
-    </div>  
-
-{% endblock content %}
