+++ template/app/static/db_models/all_models.json
@@ -1,0 +1,143 @@
+{
+    "hierarchies": {
+        "fields": {
+            "key_value": {
+                "dataType": "Text",
+                "default": false,
+                "nullable": true,
+                "relationship": null,
+                "unique": false
+            },
+            "name": {
+                "dataType": "String(256)",
+                "default": false,
+                "nullable": false,
+                "relationship": null,
+                "unique": false
+            },
+            "organisation_id": {
+                "dataType": "Integer",
+                "default": false,
+                "nullable": false,
+                "relationship": "organisations",
+                "unique": false
+            },
+            "parent_id": {
+                "dataType": "Integer",
+                "default": false,
+                "nullable": true,
+                "relationship": "hierarchies",
+                "unique": false
+            },
+            "path": {
+                "dataType": "Text",
+                "default": false,
+                "nullable": false,
+                "relationship": null,
+                "unique": false
+            },
+            "rank": {
+                "dataType": "Integer",
+                "default": false,
+                "nullable": true,
+                "relationship": null,
+                "unique": false
+            }
+        },
+        "model": "Hierarchies"
+    },
+    "organisations": {
+        "fields": {
+            "organisation_address": {
+                "dataType": "Text",
+                "default": true,
+                "nullable": true,
+                "relationship": null,
+                "unique": true
+            },
+            "organisation_city": {
+                "dataType": "String(256)",
+                "default": true,
+                "nullable": true,
+                "relationship": null,
+                "unique": true
+            },
+            "organisation_contact_email": {
+                "dataType": "String(256)",
+                "default": true,
+                "nullable": true,
+                "relationship": null,
+                "unique": true
+            },
+            "organisation_contact_name": {
+                "dataType": "String(256)",
+                "default": true,
+                "nullable": true,
+                "relationship": null,
+                "unique": true
+            },
+            "organisation_country": {
+                "dataType": "String(256)",
+                "default": true,
+                "nullable": true,
+                "relationship": null,
+                "unique": true
+            },
+            "organisation_details": {
+                "dataType": "Text",
+                "default": true,
+                "nullable": true,
+                "relationship": null,
+                "unique": true
+            },
+            "organisation_homepage": {
+                "dataType": "Text",
+                "default": true,
+                "nullable": true,
+                "relationship": null,
+                "unique": true
+            },
+            "organisation_name": {
+                "dataType": "String(256)",
+                "default": true,
+                "nullable": true,
+                "relationship": null,
+                "unique": true
+            },
+            "organisation_postal_code": {
+                "dataType": "String(256)",
+                "default": true,
+                "nullable": true,
+                "relationship": null,
+                "unique": true
+            },
+            "organisation_reg_number": {
+                "dataType": "Text",
+                "default": true,
+                "nullable": true,
+                "relationship": null,
+                "unique": true
+            },
+            "organisation_vat_number": {
+                "dataType": "Text",
+                "default": true,
+                "nullable": true,
+                "relationship": null,
+                "unique": true
+            }
+        },
+        "model": "Organisations"
+    },
+    "test": {
+        "fields": {
+            "name": {
+                "dataType": "String(256)",
+                "default": false,
+                "nullable": false,
+                "relationship": null,
+                "unique": false
+            }
+        },
+        "model": "Test"
+    }
+}
